src/third_party/blink/web_tests/svg/custom/rect-negative-corner-radii-expected.txt

CONSOLE ERROR: line 3: Error: <rect> attribute rx: A negative value is not valid. ("-50")
CONSOLE ERROR: line 3: Error: <rect> attribute ry: A negative value is not valid. ("-50")
CONSOLE ERROR: line 4: Error: <rect> attribute rx: A negative value is not valid. ("-100")
CONSOLE ERROR: line 5: Error: <rect> attribute ry: A negative value is not valid. ("-100")
